refactor(dashboard): split entities/agent.ts into agent/ directory
Separates pure Zod schemas from helper/derived modules. Public API
unchanged via the agent/index.ts barrel; entities/index.ts still
re-exports ./agent so all @/entities importers work without changes.

- schema.ts: sentinels + all Zod schemas + inferred types
- platform.ts: api-server/webhook config readers + PlatformId +
  derivePlatformAvailability + endpointsFor
- toolsets.ts: ToolsetId + deriveToolsetAvailability + meta
