+++ apps/dashboard/src/entities/agent/index.ts
@@ -1,0 +1,3 @@
+export * from "./schema";
+export * from "./platform";
+export * from "./toolsets";
