+++ apps/dashboard/src/entities/agent/schema.ts
@@ -1,7 +1,9 @@
 import { z } from "zod";
 
-import { ConfigSchema } from "./hermes-config";
-import { EnvVarSchema } from "./shared-env-set";
+import { ConfigSchema } from "../hermes-config";
+import { EnvVarSchema } from "../shared-env-set";
+
+import { isApiServerEnabled, isWebhookEnabled } from "./platform";
 
 export const ENV_SECRET_SENTINEL = "<secret>";
 
@@ -518,460 +520,3 @@ export const AgentSchema = AgentInputObjectSchema.extend({
 }).readonly();
 
 export type Agent = z.infer<typeof AgentSchema>;
-
-// API server settings are configured exclusively via env vars (see
-// docs/hermes-config/api-server.md). These helpers read those env vars off an
-// `AgentInput`.
-const API_SERVER_DEFAULT_PORT = 8642;
-
-export function isApiServerEnabled(input: AgentInput): boolean {
-  return (
-    input.env?.some(
-      (v) => v.name === "API_SERVER_ENABLED" && v.value.toLowerCase() === "true"
-    ) ?? false
-  );
-}
-
-export function getApiServerPort(input: AgentInput): number {
-  const raw = input.env?.find((v) => v.name === "API_SERVER_PORT")?.value;
-  if (raw === undefined || raw.trim() === "") return API_SERVER_DEFAULT_PORT;
-  const n = Number(raw);
-  return Number.isInteger(n) && n > 0 ? n : API_SERVER_DEFAULT_PORT;
-}
-
-// Webhook settings can be configured via config.yaml
-// (config.platforms.webhook.enabled / extra.port) or via env vars
-// (WEBHOOK_ENABLED / WEBHOOK_PORT). config.yaml is preferred and takes
-// precedence over the env vars; the env vars act as a fallback when the
-// corresponding config field is absent.
-const WEBHOOK_DEFAULT_PORT = 8644;
-
-export function isWebhookEnabled(input: AgentInput): boolean {
-  const configEnabled = input.config?.platforms?.webhook?.enabled;
-  if (configEnabled !== undefined) return configEnabled;
-  return (
-    input.env?.some(
-      (v) => v.name === "WEBHOOK_ENABLED" && v.value.toLowerCase() === "true",
-    ) ?? false
-  );
-}
-
-export function getWebhookPort(input: AgentInput): number {
-  const configPort = input.config?.platforms?.webhook?.extra?.port;
-  if (configPort !== undefined) return configPort;
-  const envRaw = input.env?.find((v) => v.name === "WEBHOOK_PORT")?.value;
-  if (envRaw !== undefined && envRaw.trim() !== "") {
-    const n = Number(envRaw);
-    return Number.isInteger(n) && n > 0 ? n : WEBHOOK_DEFAULT_PORT;
-  }
-  return WEBHOOK_DEFAULT_PORT;
-}
-
-// Message-platform availability — a derived, read-only view of which inbound
-// message platforms are wired up for an agent, computed from its config + env.
-// Not persisted.
-// https://hermes-agent.nousresearch.com/docs/user-guide/messaging
-
-export enum PlatformId {
-  ApiServer = "api-server",
-  Webhook = "webhook",
-  Slack = "slack",
-}
-
-export interface PlatformAvailability {
-  status: "available" | "unavailable";
-  /** Short explanation shown when status is not "available". */
-  reason?: string;
-  /** Listening port for HTTP platforms (api-server, webhook). */
-  port?: number;
-  /** Home channel for chat platforms (slack only), if configured. */
-  home?: string;
-  /**
-   * Fully-qualified endpoint URLs for this platform (base + subpath).
-   * For ingress endpoints the base carries no port (routing is by path).
-   * For internal (`*.svc.cluster.local`) endpoints the per-platform port is
-   * inserted before the subpath. Absent when the platform has no inbound HTTP
-   * surface (e.g. Slack Socket Mode) or when `agent.endpoint` is null.
-   */
-  endpoints?: string[];
-}
-
-/**
- * Ingress subpaths per platform, ordered by display priority.
- * `/health` is intentionally excluded — it is an infra health probe, not a
- * messaging endpoint users interact with. Single source of truth for both
- * the UI (`derivePlatformAvailability`) and the ingress builder
- * (`server/infras/kubernetes/client.ts`).
- */
-export const PLATFORM_INGRESS_SUBPATHS: Partial<Record<PlatformId, string[]>> = {
-  // /v1 is the main OpenAI-compatible path; /api is the generic alias.
-  [PlatformId.ApiServer]: ["/v1", "/api"],
-  [PlatformId.Webhook]: ["/webhooks"],
-  // Slack uses Socket Mode — no inbound HTTP subpath.
-};
-
-interface PlatformMeta {
-  label: string;
-  description: string;
-}
-
-const PLATFORM_META: Record<PlatformId, PlatformMeta> = {
-  [PlatformId.ApiServer]: {
-    label: "API Server",
-    description: "OpenAI-compatible HTTP endpoint for frontends like Open WebUI.",
-  },
-  [PlatformId.Webhook]: {
-    label: "Webhook",
-    description: "HTTP server that accepts signed webhooks and routes them to the agent.",
-  },
-  [PlatformId.Slack]: {
-    label: "Slack",
-    description: "Slack bot relayed through the gateway (Socket Mode).",
-  },
-};
-
-export function getPlatformLabel(id: PlatformId): string {
-  return PLATFORM_META[id].label;
-}
-
-export function getPlatformDescription(id: PlatformId): string {
-  return PLATFORM_META[id].description;
-}
-
-/** Ordered platform list for UI rendering. */
-export const PLATFORM_IDS: readonly PlatformId[] = [
-  PlatformId.ApiServer,
-  PlatformId.Webhook,
-  PlatformId.Slack,
-];
-
-const SLACK_REQUIRED_ENV_VARS = ["SLACK_BOT_TOKEN", "SLACK_APP_TOKEN", "SLACK_ALLOWED_USERS"];
-
-export function derivePlatformAvailability(id: PlatformId, agent: Agent): PlatformAvailability {
-  const env = agent.env ?? [];
-  const subpaths = PLATFORM_INGRESS_SUBPATHS[id];
-
-  switch (id) {
-    case PlatformId.ApiServer: {
-      if (!isApiServerEnabled(agent)) {
-        return { status: "unavailable", reason: "Set API_SERVER_ENABLED=true to enable." };
-      }
-      const port = getApiServerPort(agent);
-      return {
-        status: "available",
-        port,
-        ...endpointsFor(agent.endpoint, subpaths, port),
-      };
-    }
-    case PlatformId.Webhook: {
-      if (!isWebhookEnabled(agent)) {
-        return {
-          status: "unavailable",
-          reason: "Set WEBHOOK_ENABLED=true (or config.platforms.webhook.enabled).",
-        };
-      }
-      const port = getWebhookPort(agent);
-      return {
-        status: "available",
-        port,
-        ...endpointsFor(agent.endpoint, subpaths, port),
-      };
-    }
-    case PlatformId.Slack: {
-      const isSet = (name: string) =>
-        env.some((v) => v.name === name && v.value.trim() !== "");
-      const missing = SLACK_REQUIRED_ENV_VARS.filter((name) => !isSet(name));
-      if (missing.length > 0) {
-        return {
-          status: "unavailable",
-          reason: `Missing env var${missing.length > 1 ? "s" : ""}: ${missing.join(", ")}.`,
-        };
-      }
-      const home = env.find((v) => v.name === "SLACK_HOME_CHANNEL")?.value;
-      const homeName = env.find((v) => v.name === "SLACK_HOME_CHANNEL_NAME")?.value;
-      if (home) {
-        return { status: "available", home: homeName ? `${home} (${homeName})` : home };
-      }
-      return { status: "available" };
-    }
-  }
-}
-
-/**
- * Spread helper: returns `{ endpoints }` when both a base endpoint and at
- * least one subpath are present, otherwise `{}` (so the field stays absent).
- *
- * Internal endpoints (base containing `.svc.cluster.local`) need the
- * per-platform port inserted before the subpath, since each platform listens
- * on its own port. Ingress endpoints route by path on a single host, so no
- * port is inserted.
- */
-function endpointsFor(
-  endpoint: string | null | undefined,
-  subpaths: readonly string[] | undefined,
-  port: number | undefined,
-): { endpoints?: string[] } {
-  if (!endpoint || !subpaths || subpaths.length === 0) return {};
-  if (endpoint.includes(".svc.cluster.local")) {
-    if (port === undefined) return {};
-    return { endpoints: subpaths.map((p) => `${endpoint}:${port}${p}`) };
-  }
-  return { endpoints: subpaths.map((p) => `${endpoint}${p}`) };
-}
-
-// Toolset availability — a derived, read-only view of which Hermes toolsets
-// are usable for an agent, computed from its config + env. Not persisted.
-// https://hermes-agent.nousresearch.com/docs/user-guide/features/tools
-//
-// The catalog mirrors Hermes' CONFIGURABLE_TOOLSETS
-// (vendor/hermes-agent/hermes_cli/tools_config.py): the 25 built-in toolsets
-// in the same order. Plugin toolsets and platform-native toolsets outside
-// CONFIGURABLE_TOOLSETS are not listed here.
-
-export enum ToolsetId {
-  Web = "web",
-  Browser = "browser",
-  Terminal = "terminal",
-  File = "file",
-  CodeExecution = "codeExecution",
-  Vision = "vision",
-  Video = "video",
-  ImageGen = "imageGen",
-  VideoGen = "videoGen",
-  XSearch = "xSearch",
-  Tts = "tts",
-  Skills = "skills",
-  Todo = "todo",
-  Memory = "memory",
-  ContextEngine = "contextEngine",
-  SessionSearch = "sessionSearch",
-  Clarify = "clarify",
-  Delegation = "delegation",
-  Cronjob = "cronjob",
-  HomeAssistant = "homeAssistant",
-  Spotify = "spotify",
-  Discord = "discord",
-  DiscordAdmin = "discordAdmin",
-  Yuanbao = "yuanbao",
-  ComputerUse = "computerUse",
-}
-
-export type ToolsetStatus = "available" | "unavailable";
-
-export interface ToolsetAvailability {
-  status: ToolsetStatus;
-  /** Short explanation shown when status is not "available". */
-  reason?: string;
-}
-
-interface ToolsetMeta {
-  label: string;
-  description: string;
-}
-
-const TOOLSET_META: Record<ToolsetId, ToolsetMeta> = {
-  [ToolsetId.Web]: {
-    label: "Web",
-    description: "Search the web and extract page content.",
-  },
-  [ToolsetId.Browser]: {
-    label: "Browser",
-    description: "Interactive browser automation.",
-  },
-  [ToolsetId.Terminal]: {
-    label: "Terminal",
-    description: "Execute shell commands.",
-  },
-  [ToolsetId.File]: {
-    label: "File",
-    description: "Read, edit, and search files.",
-  },
-  [ToolsetId.CodeExecution]: {
-    label: "Code Execution",
-    description: "Run code in an isolated sandbox.",
-  },
-  [ToolsetId.Vision]: {
-    label: "Vision",
-    description: "Analyze images with a vision-capable model.",
-  },
-  [ToolsetId.Video]: {
-    label: "Video",
-    description: "Analyze video with a video-capable model.",
-  },
-  [ToolsetId.ImageGen]: {
-    label: "Image Generation",
-    description: "Generate images from text prompts.",
-  },
-  [ToolsetId.VideoGen]: {
-    label: "Video Generation",
-    description: "Generate video from text, image, or reference input.",
-  },
-  [ToolsetId.XSearch]: {
-    label: "X Search",
-    description: "Search X (Twitter) posts via xAI.",
-  },
-  [ToolsetId.Tts]: {
-    label: "Text-to-Speech",
-    description: "Convert text to spoken audio.",
-  },
-  [ToolsetId.Skills]: {
-    label: "Skills",
-    description: "List, view, and manage installed skills.",
-  },
-  [ToolsetId.Todo]: {
-    label: "Todo",
-    description: "Track and plan tasks within a conversation.",
-  },
-  [ToolsetId.Memory]: {
-    label: "Memory",
-    description: "Persistent memory and recall across sessions.",
-  },
-  [ToolsetId.ContextEngine]: {
-    label: "Context Engine",
-    description: "Runtime tools from the active context engine.",
-  },
-  [ToolsetId.SessionSearch]: {
-    label: "Session Search",
-    description: "Search past conversation sessions.",
-  },
-  [ToolsetId.Clarify]: {
-    label: "Clarify",
-    description: "Ask the user clarifying questions.",
-  },
-  [ToolsetId.Delegation]: {
-    label: "Delegation",
-    description: "Delegate tasks to subagents.",
-  },
-  [ToolsetId.Cronjob]: {
-    label: "Cron",
-    description: "Schedule and manage recurring jobs.",
-  },
-  [ToolsetId.HomeAssistant]: {
-    label: "Home Assistant",
-    description: "Control smart home devices via Home Assistant.",
-  },
-  [ToolsetId.Spotify]: {
-    label: "Spotify",
-    description: "Control Spotify playback, playlists, and library.",
-  },
-  [ToolsetId.Discord]: {
-    label: "Discord",
-    description: "Read and participate in Discord channels.",
-  },
-  [ToolsetId.DiscordAdmin]: {
-    label: "Discord Admin",
-    description: "Administer Discord servers: channels, roles, pins.",
-  },
-  [ToolsetId.Yuanbao]: {
-    label: "Yuanbao",
-    description: "Query Yuanbao groups, members, and DMs.",
-  },
-  [ToolsetId.ComputerUse]: {
-    label: "Computer Use",
-    description: "Drive the desktop via cua-driver on macOS, Windows, or Linux.",
-  },
-};
-
-export function getToolsetLabel(id: ToolsetId): string {
-  return TOOLSET_META[id].label;
-}
-
-export function getToolsetDescription(id: ToolsetId): string {
-  return TOOLSET_META[id].description;
-}
-
-/** Ordered toolset list for UI rendering. Mirrors CONFIGURABLE_TOOLSETS order. */
-export const TOOLSET_IDS: readonly ToolsetId[] = [
-  ToolsetId.Web,
-  ToolsetId.Browser,
-  ToolsetId.Terminal,
-  ToolsetId.File,
-  ToolsetId.CodeExecution,
-  ToolsetId.Vision,
-  ToolsetId.Video,
-  ToolsetId.ImageGen,
-  ToolsetId.VideoGen,
-  ToolsetId.XSearch,
-  ToolsetId.Tts,
-  ToolsetId.Skills,
-  ToolsetId.Todo,
-  ToolsetId.Memory,
-  ToolsetId.ContextEngine,
-  ToolsetId.SessionSearch,
-  ToolsetId.Clarify,
-  ToolsetId.Delegation,
-  ToolsetId.Cronjob,
-  ToolsetId.HomeAssistant,
-  ToolsetId.Spotify,
-  ToolsetId.Discord,
-  ToolsetId.DiscordAdmin,
-  ToolsetId.Yuanbao,
-  ToolsetId.ComputerUse,
-];
-
-export function deriveToolsetAvailability(id: ToolsetId, agent: Agent): ToolsetAvailability {
-  const config = agent.config;
-  const env = agent.env ?? [];
-  const hasEnv = (name: string) => env.some((v) => v.name === name && v.value.trim() !== "");
-
-  switch (id) {
-    case ToolsetId.Web: {
-      const hasBackend = !!(
-        config?.web?.backend ??
-        config?.web?.search_backend ??
-        config?.web?.extract_backend
-      );
-      return hasBackend
-        ? { status: "available" }
-        : { status: "unavailable", reason: "No web backend configured." };
-    }
-    case ToolsetId.XSearch: {
-      // Gated on xAI credentials; off by default.
-      const hasXaiKey = hasEnv("XAI_API_KEY");
-      const xaiBackend = config?.web?.search_backend === "xai";
-      return hasXaiKey || xaiBackend
-        ? { status: "available" }
-        : { status: "unavailable", reason: "Set XAI_API_KEY to opt in." };
-    }
-    case ToolsetId.Browser: {
-      return config?.browser?.cloud_provider
-        ? { status: "available" }
-        : { status: "unavailable", reason: "No browser provider configured." };
-    }
-    case ToolsetId.HomeAssistant: {
-      return hasEnv("HASS_TOKEN")
-        ? { status: "available" }
-        : { status: "unavailable", reason: "Set HASS_TOKEN to enable Home Assistant." };
-    }
-    case ToolsetId.Discord:
-    case ToolsetId.DiscordAdmin: {
-      return hasEnv("DISCORD_BOT_TOKEN")
-        ? { status: "available" }
-        : { status: "unavailable", reason: "Set DISCORD_BOT_TOKEN to enable Discord." };
-    }
-    case ToolsetId.ImageGen: {
-      const hasFalKey = hasEnv("FAL_KEY");
-      const useGateway = config?.image_gen?.use_gateway === true;
-      return hasFalKey || useGateway
-        ? { status: "available" }
-        : { status: "unavailable", reason: "Set FAL_KEY or enable the managed gateway." };
-    }
-    case ToolsetId.VideoGen: {
-      const hasFalKey = hasEnv("FAL_KEY");
-      const hasXaiKey = hasEnv("XAI_API_KEY");
-      return hasFalKey || hasXaiKey
-        ? { status: "available" }
-        : { status: "unavailable", reason: "Set FAL_KEY or XAI_API_KEY for video generation." };
-    }
-    case ToolsetId.Spotify:
-      return { status: "unavailable", reason: "Spotify OAuth is not supported in Hermeum." };
-    case ToolsetId.ComputerUse:
-      return { status: "unavailable", reason: "Computer Use is not supported in Hermeum." };
-    case ToolsetId.Yuanbao:
-      return { status: "unavailable", reason: "Only available on the Yuanbao messaging platform." };
-    default:
-      // Toolsets with no config gate in Hermeum. They are available whenever
-      // the agent exists; per-agent toggling is a follow-up.
-      return { status: "available" };
-  }
-}
